Add tests for player and gameboard
